+++ src/category.ts
@@ -3,7 +3,7 @@ import { I18nString } from "./common";
 export interface Category {
   id: number;
   name: I18nString;
-  treeNodeLevel: number;
-  productsCount: number;
   parentId: number;
+  productsCount: number;
+  treeNodeLevel: number;
 }
