+++ src/common.ts
@@ -6,15 +6,15 @@ export interface I18nString {
 }
 
 export interface Address {
-  first_name: string;
-  last_name: string;
-  company: string;
   address_1: string;
   address_2: string;
   city: string;
-  state: string;
-  postcode: string;
+  company: string;
   country: string;
   email: string;
+  first_name: string;
+  last_name: string;
   phone: string;
+  postcode: string;
+  state: string;
 }
