+++ src/coupon.ts
@@ -1,28 +1,28 @@
 export interface Coupon {
-  code: string;
-  discount: {
-    value: number;
-    type: "%" | "$";
-  };
-  total: {
-    value: number;
-    type: "%" | "$";
-  };
-  shipping: {
-    value: number;
-    type: "%" | "$";
-  };
-  tax: {
-    value: number;
-    type: "%" | "$";
-  };
-  type: "salesorder" | "subscription";
-  startDate: Date;
-  endDate: Date;
-  maxUses: number;
-  minAppliedAmount: number;
   appliedMemberships: string[];
   auto: boolean;
   campaignName: string;
+  code: string;
+  discount: {
+    type: "%" | "$";
+    value: number;
+  };
+  endDate: Date;
+  maxUses: number;
+  minAppliedAmount: number;
+  shipping: {
+    type: "%" | "$";
+    value: number;
+  };
+  startDate: Date;
+  tax: {
+    type: "%" | "$";
+    value: number;
+  };
+  total: {
+    type: "%" | "$";
+    value: number;
+  };
+  type: "salesorder" | "subscription";
   useCount: number;
 }
