+++ src/invoice.ts
@@ -1,18 +1,18 @@
 export interface Invoice {
-  invoice_id: string;
-  customer_name: string;
+  adjustment: number;
+  balance: number;
+  coupon: string;
+  created_time: Date;
   customer_id: string;
-  status: string;
-  invoice_number: string;
-  reference_number: string;
+  customer_name: string;
   date: Date;
   due_date: Date;
   due_days: number;
-  total: number;
-  balance: number;
-  created_time: Date;
+  invoice_id: string;
+  invoice_number: string;
   last_modified_time: Date;
+  reference_number: string;
   shipping_charge: number;
-  adjustment: number;
-  coupon: string;
+  status: string;
+  total: number;
 }
