+++ src/shipping-policy.ts
@@ -1,22 +1,22 @@
 export interface ShipmentPolicy {
-  name: string;
   countries: string[];
+  name: string;
   rules: [
     {
-      courier: string;
-      delivery_days_min: number;
-      delivery_days_max: number;
-      units_min: number;
-      units_max: number;
-      type: "weight" | "price";
       cost: number;
+      courier: string;
+      delivery_days_max: number;
+      delivery_days_min: number;
+      type: "weight" | "price";
+      units_max: number;
+      units_min: number;
     }
   ];
   ship_from: [
     {
+      Enum: string;
       city: string;
       country: string;
-      Enum: string;
     }
   ];
 }
