--- webhook.ts
@@ -1,10 +1,0 @@
-export interface Webhook {
-    id: string;
-    // TODO: enum topics
-    topic: string;
-    supplierId: string;
-    url: string;
-    createdAt: Date;
-    updatedAt: Date;
-  }
-
